Converting most types to floats and sequencing Drive better.

+++ src/Engine.hpp
@@ -30,7 +30,7 @@ public:
     void Init(Arena arena);
     void Load(std::vector<std::shared_ptr<IRobot>>&& robots);
     void Tick();
-    uint32_t ScanResult(uint32_t robot_id, uint32_t degree, uint32_t resolution) const;
+    float ScanResult(uint32_t robot_id, uint32_t degree, uint32_t resolution) const;
     void AddShot(Shot shot);
     const Arena& GetArena() const;
     const std::vector<std::shared_ptr<IRobot>>& GetRobots() const;
